renaming client file
